extract LikeCommentShare
