+++ src/components/Comments/Display/ByUrl/index.tsx
@@ -20,9 +20,22 @@ import * as Colors from "consts/Colors"
 import { Reaction, Comment } from "entities"
 
 // Relative
-import { GET_COMMENTS_BY_URL_QUERY } from "./query"
+import { GET_COMMENTS_BY_URL_QUERY, CommentOrderByInput } from "./query"
 import { CommentRenderer, FlexRow, FlexColumn } from "../../../../pages/rds"
 
+const sortByNewest = _.flow(_.partialRight(_.sortBy, "created"), _.reverse)
+
+/**
+ * This component takes `url` prop, and fetches all comments by this
+ * url.
+ *
+ * @usage
+ * ```tsx
+ * import { CommentsByUrl } from "components/Comments/Display/ByUrl"
+ *
+ * <CommentsByUrl url={location.pathname} />
+ * ```
+ */
 export const CommentsByUrl = ({ url }) => {
   const { currentUserId } = useAuthentication()
 
@@ -31,7 +44,9 @@ export const CommentsByUrl = ({ url }) => {
     { data, loading: queryLoading, query: queryError },
   ] = useLazyQuery<{
     getCommentsByUrl: Comment[]
-  }>(GET_COMMENTS_BY_URL_QUERY, { variables: { url: url } })
+  }>(GET_COMMENTS_BY_URL_QUERY, {
+    variables: { url: url, filter: CommentOrderByInput.created_DESC },
+  })
 
   const didIntersect = useRef(false)
   const [isIntersecting, bind] = useIO({
@@ -51,7 +66,7 @@ export const CommentsByUrl = ({ url }) => {
    */
   if (queryLoading)
     return (
-      <CommentRenderer>
+      <CommentRenderer {...bind}>
         <FlexRow style={{ marginBottom: `.5rem` }}>
           <p>Loading</p>
         </FlexRow>
@@ -59,16 +74,19 @@ export const CommentsByUrl = ({ url }) => {
     )
   if (data?.getCommentsByUrl.length < 1)
     return (
-      <CommentRenderer>
+      <CommentRenderer {...bind}>
         <FlexRow style={{ marginBottom: `.5rem` }}>
           <p>Be the first to comment!</p>
         </FlexRow>
       </CommentRenderer>
     )
+  /**
+   * @TODO still need to display reactions
+   */
   return (
     data?.getCommentsByUrl.map((_comment, i) => {
       return (
-        <CommentRenderer key={_comment.id}>
+        <CommentRenderer key={_comment.id} {...bind}>
           <FlexRow style={{ marginBottom: `.5rem` }}>
             <Avatar src={_comment.user.avatar_url} />
             <FlexColumn>
