+++ src/components/Comments/Display/ByUrl/query.ts
@@ -1,8 +1,13 @@
 import { gql } from "apollo-boost"
 
+export enum CommentOrderByInput {
+  created_ASC = "created_ASC",
+  created_DESC = "created_DESC",
+}
+
 export const GET_COMMENTS_BY_URL_QUERY = gql`
-  query($url: String!) {
-    getCommentsByUrl(url: $url) {
+  query($url: String!, $filter: CommentOrderByInput) {
+    getCommentsByUrl(url: $url, filter: $filter) {
       id
       body
       created
