+++ src/pages/rds.tsx
@@ -3,14 +3,12 @@ import * as React from "react"
 import { navigate } from "gatsby"
 import moment from "moment"
 import _ from "lodash"
-import { graphql, Link } from "gatsby"
+import { graphql } from "gatsby"
 import { useSpring, animated } from "react-spring"
 import styled, { BaseProps } from "styled-components"
 import theme from "styled-theming"
-import { useMediaQuery } from "@material-ui/core"
 import { useApolloClient, useMutation } from "@apollo/react-hooks"
 import { gql } from "apollo-boost"
-import Popover from "@material-ui/core/Popover"
 
 // Hooks
 import { useCommentLogic } from "hooks/rds/useCommentLogic"
@@ -26,10 +24,7 @@ import { LoadingIndicator } from "components/LoadingIndicator"
 import { CreateComment } from "components/Comments/Create"
 import { SubmitButton } from "components/Form"
 import { Avatar } from "components/Avatar"
-
-// Other
-import { POSSIBLE_VARIANTS } from "entities/Reaction"
-import * as Colors from "consts/Colors"
+import { LikeCommentShare } from "components/LikeCommentShare"
 
 const LEFT_OFFSET = 20
 
@@ -44,7 +39,7 @@ const backgroundPosition = theme.variants("mode", "variant", {
   None: { light: "0% 0%", dark: "0% 0%" }, // :_: | :_:
 })
 
-const Variant = styled(animated.div)`
+export const Variant = styled(animated.div)`
   transition: background-position 150ms ease-in-out;
   will-change: background;
   background-image: url(https://pf-emoji-service--cdn.us-east-1.prod.public.atl-paas.net/standard/a51a7674-8d5d-4495-a2d2-a67c090f5c3b/32x32/spritesheets/people.png);
@@ -97,58 +92,6 @@ export const CommentRenderer = styled(animated.div)`
     margin-bottom: 1rem; /* 16px */
   }
 `
-const Line = styled(animated.div)`
-  height: 1px;
-  background: ${borderColor};
-`
-
-const popoverBackground = theme("mode", {
-  // light: Colors.LIGHT_GRADIENTS[0],
-  // dark: Colors.DARK_GRADIENTS[0],
-  light: Colors.silverLighter,
-  dark: Colors.blackLighter,
-})
-const VariantButton = styled(animated.button)`
-  background: none;
-  display: flex;
-  justify-content: center;
-  border-radius: 0.2rem;
-  height: ${ITEM_HEIGHT + 10}px;
-  width: ${ITEM_HEIGHT + 10}px;
-  padding: 0px;
-  border: none;
-
-  transition: background 150ms ease-in-out;
-  :hover {
-    background: ${borderColor};
-  }
-`
-
-const FlexBoxButton = styled.div`
-  align-items: center;
-  border-radius: 0.2rem;
-  cursor: pointer;
-  display: flex;
-  flex: 1;
-  height: 3rem;
-  justify-content: center;
-  margin: 5px;
-
-  transition: background 200ms ease-in-out;
-  will-change: background;
-
-  color: ${theme("mode", {
-    light: (props: BaseProps) => props.theme.formInput.color,
-    dark: (props: BaseProps) => props.theme.formInput.color,
-  })};
-
-  :hover {
-    background: ${theme("mode", {
-      light: (props: BaseProps) => props.theme.commentRenderer.borderColor,
-      dark: (props: BaseProps) => props.theme.commentRenderer.borderColor,
-    })};
-  }
-`
 
 /**
  * @TODO
@@ -177,97 +120,6 @@ const UploadButton = styled(animated.button)`
   border: 1px solid lightgray;
   border-radius: 0.25rem;
 `
-
-const PopoverContents = styled(animated.div)`
-  color: ${theme("mode", {
-    light: (props: BaseProps) => props.theme.formInput.color,
-    dark: (props: BaseProps) => props.theme.formInput.color,
-  })};
-  background: ${popoverBackground};
-  display: flex;
-  flex-direction: row;
-  padding: 1rem;
-
-  :first-child > :not(:first-child) {
-    /* border: 1px dotted red; */
-    margin-left: 0.75rem;
-  }
-`
-const REACT_TO_COMMENT = gql`
-  mutation($variant: ReactionVariant!, $commentId: Int!) {
-    reactToComment(variant: $variant, commentId: $commentId) {
-      id
-    }
-  }
-`
-const LikeOrComment = ({ commentId }: { commentId: number }) => {
-  const { currentUserId } = useAuthentication()
-  const windowSm = useMediaQuery("(max-width:480px)")
-  const [anchorEl, setAnchorEl] = React.useState<HTMLButtonElement | null>(null)
-
-  const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-    setAnchorEl(event.currentTarget)
-  }
-
-  const handleClose = () => {
-    setAnchorEl(null)
-  }
-
-  const open = Boolean(anchorEl)
-  const id = open ? "simple-popover" : undefined
-
-  const [reactToComment, { data, loading }] = useMutation(REACT_TO_COMMENT)
-  const handleSelectReaction = ({ variant }) => event => {
-    reactToComment({
-      variables: { variant: variant, commentId },
-    })
-    handleClose()
-  }
-
-  return (
-    <>
-      <Line />
-      <FlexRow>
-        <FlexBoxButton onClick={handleClick}>React</FlexBoxButton>
-        <Popover
-          id={id}
-          open={open}
-          anchorEl={anchorEl}
-          onClose={handleClose}
-          anchorOrigin={{
-            vertical: "bottom",
-            horizontal: "center",
-          }}
-          transformOrigin={{
-            vertical: "top",
-            horizontal: "center",
-          }}
-        >
-          <PopoverContents>
-            {currentUserId ? (
-              POSSIBLE_VARIANTS.map((variant, i) => (
-                <VariantButton
-                  key={variant}
-                  onClick={handleSelectReaction({ variant: variant })}
-                >
-                  <Variant variant={variant} />
-                </VariantButton>
-              ))
-            ) : (
-              <>
-                Please&nbsp;
-                <Link to={"/auth/login"}>{"log in"}</Link>
-                &nbsp;to react
-              </>
-            )}
-          </PopoverContents>
-        </Popover>
-        <FlexBoxButton>Comment</FlexBoxButton>
-        {!windowSm && <FlexBoxButton>Share</FlexBoxButton>}
-      </FlexRow>
-    </>
-  )
-}
 
 const RdsPage = props => {
   const { currentUserId } = useAuthentication()
@@ -448,7 +300,7 @@ const RdsPage = props => {
                 </div>
               </ReactionsContainer>
             </Container>
-            <LikeOrComment commentId={parseInt(_comment.id)} />
+            <LikeCommentShare commentId={parseInt(_comment.id)} />
           </CommentRenderer>
         ))}
       </Container>
